@@ -1,38 +1,18 @@
 package com.dagf.winesupport;
 
-import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
-import android.media.AudioManager;
-import android.media.MediaPlayer;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.design.widget.TabLayout;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
-import android.view.View;
-import android.widget.EditText;
-import android.widget.TextView;
-import android.widget.Toast;
 
 import com.dagf.presentlogolib.WineHelper;
-import com.dagf.presentlogolib.eq.EqualizerFragment;
-import com.dagf.presentlogolib.fragmentssec.MusicFragment;
-import com.dagf.presentlogolib.fragmentssec.PlayerService;
 import com.dagf.presentlogolib.nextview.NextViewDagf;
-import com.dagf.presentlogolib.nextview.NextViewItem;
-import com.dagf.presentlogolib.utils.MediafireParser;
+import com.dagf.presentlogolib.utils.UpdateDialog;
 import com.dagf.presentlogolib.utils.Updater;
-
-import java.io.File;
-import java.io.FilenameFilter;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Timer;
-import java.util.TimerTask;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -76,6 +56,26 @@ if(WineHelper.checkPermissionWrite(this)) {
 
     }
 
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+
+        if(requestCode == UpdateDialog.MY_PERMISSIONS_INSTALLFROM_UNKWONSOURCES){
+            Log.e("MAIN", "onActivityResult: SI ES ACT");
+            try {
+                Updater.check(this, "https://moviesapp.website/secure/cpanels/admin-movie-tv/api/");
+
+            } catch (PackageManager.NameNotFoundException e) {
+                Log.e("MAIN", "onCreate: "+e.getMessage() );
+                e.printStackTrace();
+            }
+        }
+
+        if(requestCode == UpdateDialog.key_unnistall){
+            Updater.NowOpenAPK();
+        }
+
+    }
 
     @Override
     public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
