@@ -5,7 +5,7 @@ import android.media.MediaMetadataRetriever;
 import android.os.Build;
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.Log;
 
 import java.util.HashMap;
